Changed some minor design issues

+++ src/pages/Home/index.js
@@ -43,7 +43,7 @@ const Home = props => {
     const lowerBound = ratingFilter * 2;
     const upperBound = ratingFilter * 2 + 2;  
     const filteredMovies =   movieList.filter(movie=>{
-       const avgVote = parseInt(movie.vote_average);
+       const avgVote = parseFloat(movie.vote_average);
       if( avgVote >= lowerBound && avgVote <=upperBound ){
           return movie;
        }
